requirements.txt
PyGithub
python-dotenv
packaging
python-gitlab
python-dateutil
requests
urllib3